test(coach-ops): bind UNIT-003 card-presentation to its own test file
Split card-internal rendering tests out of test_dashboard_unit_002_render_grid
into test_dashboard_unit_003_card_presentation so acc:coach-ops:M002-UNIT-003
has an anchored test (bidirectional-binding validator, #410). Rename the filter
file 003→004 to match its UNIT-004 header.
